refactoring out section header view
